backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
Flasgger
python-dotenv
pymysql
flask-cors